+++ rbdc-mysql/src/types/timestamp.rs
@@ -63,7 +63,7 @@ impl Decode for Timestamp {
                         day: date.day,
                         mon: date.mon,
                         year: date.year,
-                        offset: fastdate::offset_sec() as i64,
+                        offset: fastdate::offset_sec(),
                     }
                     .unix_timestamp_millis() as u64,
                 )
