+++ rbdc-pg/src/types/time.rs
@@ -23,7 +23,7 @@ impl Decode for Time {
                     day: 0,
                     mon: 0,
                     year: 0,
-                    offset: fastdate::offset_sec() as i64,
+                    offset: fastdate::offset_sec(),
                 };
                 let t = {
                     if us < 0 {
